move clap_timestamp and clap_universal_plugin_id into separate files to match upstream
