+++ src/timestamp.rs
@@ -1,0 +1,3 @@
+pub type clap_timestamp = u64;
+
+pub const CLAP_TIMESTAMP_UNKNOWN: clap_timestamp = 0;
